FE: Solve cookie notice bar bug & make prettier
The cookies notice bar was blocked by the navigation bar on phones. Reason being the cookie bars height was based on a percentage of the body. With the navigation bar on phones counting to the height of the body, it stops working as intended as the navigation bar steals from the total height.
Solution: Make the height of cookies notice bar in fixed pixels for phones.