Solve Roslyn not found error
This solves this error message: "Could not find a part of the path 'C:\Dev\Projects\projectname\bin\roslyn\csc.exe'." which is created when Roslyn is removed from /packages folder.
Solution: Run this line in package manager console for the web project: Microsoft.CodeDom.Providers.DotNetCompilerPlatform -r
It will reinstall that package which uses Roslyn.
From stackoverflow thread: https://stackoverflow.com/questions/32780315/could-not-find-a-part-of-the-path-bin-roslyn-csc-exe