Fix Umbraco extension method GetCropUrl exception
An exception is thrown when the method is called in unit tests.
Error message: "Argument out of range exception. Not a valid Win32 FileTime"

Cause: The extension method uses ToFileTimeUtc() method on the media items property "UpdateDate". By default in unit tests, the property value is DateTime.Min.
Solution: Mock the property UpdateDate on the media item to be, for example. DateTime.Now()